Refactor TransactionEntryList for improved data handling and structure
- Remove the Data and Pending components, consolidating their functionality directly into TransactionEntryList.
- Update TransactionEntryList to handle collection fetching and container composition, enhancing clarity and performance.
- Adjust README to reflect changes in transaction-entry component structure and responsibilities.
